% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink_correction.R
\name{blink_correction}
\alias{blink_correction}
\title{blink correction}
\usage{
blink_correction(
  signal,
  lower_blink_range = 75,
  upper_blink_range = 250,
  cut_blink_data = 25,
  sampling_rate = 300
)
}
\arguments{
\item{signal}{signal a numeric vector, typically of raw pupillometry data.}

\item{lower_blink_range}{lower bound of blinks in ms. Default 25ms.}

\item{upper_blink_range}{upper bound of blinks in ms. Default 250ms.}

\item{cut_blink_data}{data to cut before and after blinks in ms. Default 25ms}

\item{sampling_rate}{eye-tracker sampling rate in Hz. Default 300Hz}
}
\value{
signal cleared of potential blink data. These are set to NA.
}
\description{
This function corrects a numeric vector of pupil data of potential eye blinks.
}
\examples{
to be done
}
