% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_invalid.R
\name{exclude_invalid}
\alias{exclude_invalid}
\title{exclude invalid data}
\usage{
exclude_invalid(
  pupil_diameter,
  lower_bound_pupil_diameter = 2,
  upper_bound_pupil_diameter = 8,
  check_raw_validity = FALSE,
  validity_data,
  validity_threshold = 2
)
}
\arguments{
\item{pupil_diameter}{vector of pupil data as numeric.}

\item{lower_bound_pupil_diameter}{lower bound of plausible pupil diameter.}

\item{upper_bound_pupil_diameter}{upper bound of plausible pupil diameter.}

\item{check_raw_validity}{TRUE versus FALSE to check pupil data validity based on external validity information.}

\item{validity_data}{vector of length pupil diamter that contains external data validity information as integer.}

\item{validity_threshold}{integer as threshold to validate pupil data.}
}
\value{
pupil data corrected for invalid data that is outside plausible range (optional: and above validity threshold).
}
\description{
This function exclude invalid pupillometry data.
}
\examples{
#to be done
}
