% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_size_outlier.R
\name{exclude_size_outlier}
\alias{exclude_size_outlier}
\title{exclude size outlier}
\usage{
exclude_size_outlier(signal, timestamp, MAD_constant = 3, smooth_length = 150)
}
\arguments{
\item{signal}{signal as a numeric vector, typically of raw pupillometry data.}

\item{timestamp}{timestamp in the raw data associated with signal. Expects millisecond format timestamps}

\item{MAD_constant}{constant as threshold to be used to define outlier. Default 3.}

\item{smooth_length}{length of linear imputation to be applied before outlier exclusion}
}
\value{
signal cleared of speed outliers. These are set to NA.
}
\description{
This function estimates the median absolute deviation of pupil size from a trend line.
Applies a two pass approach: first pass excludes deviation from trend line derived from all samples.
second pass excludes deviation from trend line derived from samples passing first pass.
reintroduction of samples that might have been falsely excluded due to outliers.
Detected size outlier are set to NA.
}
\examples{
#to be done
}
