% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_speed_outlier.R
\name{exclude_speed_outlier}
\alias{exclude_speed_outlier}
\title{exclude speed outlier}
\usage{
exclude_speed_outlier(signal, timestamp, MAD_constant = 3)
}
\arguments{
\item{signal}{signal as a numeric vector, typically of raw pupillometry data.}

\item{timestamp}{timestamp in the raw data associated with signal. Works independently of the resolution of the timestamp}

\item{MAD_constant}{constant as threshold to be used to define outlier. Default 3.}
}
\value{
signal cleared of speed outliers. These are set to NA.
}
\description{
This function estimates speed of pupil size change based on timestamp data
and excludes data that is outside defined ranges. Detected speed outlier are set to NA.
Consider time jumps in timestamp vector. Speed change is estimate forward and backwards.
}
\examples{
#to be done
}
