% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_interpolation.R
\name{missing_interpolation}
\alias{missing_interpolation}
\title{missing interpolation}
\usage{
missing_interpolation(signal, timestamp, interpolation_length = 300)
}
\arguments{
\item{signal}{numeric vector of pupil size data.}

\item{timestamp}{timestamp in the raw data associated with signal. Expects millisecond format timestamps.}

\item{interpolation_length}{gaps of missing data to interpolate in milliseconds}
}
\value{
pupil data with missing interpolated.
}
\description{
missing data as NA is linearily interpolated for a specific timeframe
}
\examples{
#to be done
}
