% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset_interpolation.R
\name{offset_interpolation}
\alias{offset_interpolation}
\title{offset interpolation}
\usage{
offset_interpolation(left_pupil_data, right_pupil_data, which_eye)
}
\arguments{
\item{left_pupil_data}{pupil size data of left eye.}

\item{right_pupil_data}{pupil size data of right eye.}

\item{which_eye}{data of which eye to interpolate by offset: "left" or "right"}
}
\value{
pupil data with missing interpolated for offset.
}
\description{
when pupil size data is provided for both eye. The function estimates the
size offset between both eyes and interpolates missing data in one eye by
size estimates of the other eye plus offset
}
\examples{
#to be done
}
