% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_preprocessing.R
\name{pupil_preprocessing}
\alias{pupil_preprocessing}
\title{pupil preprocessing}
\usage{
pupil_preprocessing(
  eye_tracking_raw,
  provide_variable_names = F,
  left_diameter_name = NA,
  right_diameter_name = NA,
  timestamp_name = NA,
  diagnostic_output = T,
  lower_bound_pupil_diameter = 2,
  upper_bound_pupil_diameter = 8,
  check_raw_validity = F,
  validity_threshold = 2,
  lower_blink_range = 75,
  upper_blink_range = 250,
  cut_blink_data = 25,
  sampling_rate = NA,
  MAD_constant = 3,
  smooth_length = 150,
  NA_gap_length = 75,
  sparsity_filter_island_length = 50,
  interpolation_length = 300,
  timestamp_scaling = 1
)
}
\arguments{
\item{eye_tracking_raw}{raw pupil data with estimates for both eyes, timestamps, and optional validity estimates. Expects the following variable names: left_pupil_diameter, right_pupil_diameter, timestamp. Otherwise provide names with provide_variable_names option.}

\item{provide_variable_names}{provide variable names as string for left_diameter, right_diameter, timestamp as provided by the raw data stream.}

\item{left_diameter_name}{if provide_variable_names=T, provide name as string}

\item{right_diameter_name}{if provide_variable_names=T, provide name as string}

\item{timestamp_name}{if provide_variable_names=T, provide name as string. timestamp data is expected in millisecond format.}

\item{diagnostic_output}{provide diagnostic output that shows excluded data per preprocessing step.}

\item{lower_bound_pupil_diameter}{smallest pupil size that is considered valid in mm, default = 2.}

\item{upper_bound_pupil_diameter}{lergest pupil size that is considered valid in mm, default = 8.}

\item{check_raw_validity}{consider eye tracker validity estimates in excluding invalid data.}

\item{validity_threshold}{threshold of validity estimates that are considered valid, set between 1 (aggressive) to 4 (very conservative) - specific to TOBII TX-300, default = 2.}

\item{lower_blink_range}{shortest duration that is considered a blink in milliseconds, default = 75.}

\item{upper_blink_range}{longest duration that is considered a blink in milliseconds, default = 250.}

\item{cut_blink_data}{data to cut before and after blinks in ms, default 25ms.}

\item{sampling_rate}{eye-tracker sampling rate in Hz, is estimated from timetamps by default, but can be provided explicitly here}

\item{MAD_constant}{Mean Average Deviation constant as threshold of deviation from mean to be used to define outlier. Default 3.}

\item{smooth_length}{length of linear imputation to be applied before outlier exclusion.}

\item{NA_gap_length}{gaps >= this length in milliseconds are considered "big" and define split points in the sparsity filtering.}

\item{sparsity_filter_island_length}{valid sections shorter than this in ms are rejected and thus set to NA.}

\item{interpolation_length}{gaps shorter than this in millisecond that will be linearly interpolated.}

\item{timestamp_scaling}{numeric. Timestamp data in eye-tracking_raw is expected in millisecond format. the timestamp_scaling allows to adjust, if different timetamp resolutions are provided: second format: timestamp_scaling = 1000, microsecond format: timestamp_scaling = 0.001, default is 1 as expected millisecond format.}
}
\value{
returns preprocessed pupil size estimate as mean of both eyes.
}
\description{
wrapper function that takes pupil dilation raw data of both eye and preprocesses into a cleaned pupil size of both eye
Includes the following steps that are each available as separate functions:
\enumerate{
\item exclude invalid data - exclude_invalid()
\item exclude data around blinks - blink_correction()
\item exclude dilation speed outlier - exclude_speed_outlier()
\item exclude dilation size outlier - exclude_size_outlier()
\item sparsity filter - exclude islands of data around large gaps of missings
\item interpolate by eye offset - offset_interpolation()
\item interpolate gaps of missing data - missing_interpolation()
\item calculate mean of both eyes
}
}
\examples{
#to be done
}
