% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PupilPreprocess-package.R
\docType{package}
\name{PupilPreprocess-package}
\alias{PupilPreprocess}
\alias{PupilPreprocess-package}
\title{PupilPreprocess: Pupil data preprocessing}
\description{
preprocessing of pupil data from eye-tracker data streams.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicobast/PupilPreprocess}
  \item Report bugs at \url{https://github.com/nicobast/PupilPreprocess/issues}
}

}
\author{
\strong{Maintainer}: Nico Bast \email{nbast@med.uni-frankfurt.de}

}
\keyword{internal}
