% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsity_filtering.R
\name{sparsity_filtering}
\alias{sparsity_filtering}
\title{sparsity filtering}
\usage{
sparsity_filtering(signal, timestamp, gap_ms = 75, min_segment_ms = 50)
}
\arguments{
\item{signal}{numeric vector with NAs where data are invalid.}

\item{timestamp}{timestamp in the raw data associated with signal. Expects millisecond format timestamps.}

\item{gap_ms}{gaps >= this length in ms are considered "big" and define split points.}

\item{min_segment_ms}{valid sections shorter than this in ms are rejected and thus set to NA}
}
\value{
signal cleared of sparsity islands. These are set to NA.
}
\description{
This function removes short islands of valid samples that are bounded by large gaps.
}
\examples{
#to be done
}
