% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pupil_data.R
\name{generate_pupil_data}
\alias{generate_pupil_data}
\title{generate_pupil_data}
\usage{
generate_pupil_data(
  sampling_rate = 300,
  phase_duration = 1,
  number_of_responses = 100,
  mean_response_magnitude = 0.05,
  introduce_NA = T,
  NA_proportion = 0.2,
  introduce_noise = T,
  noise_magnitude = 0.3,
  start_value_pupil_size = 4,
  drift = -0.002
)
}
\arguments{
\item{sampling_rate}{sampling rate of the sampled data.}

\item{phase_duration}{pupil response duration in seconds.}

\item{number_of_responses}{number of pupil responses to sample.}

\item{mean_response_magnitude}{mean pupil response magnitude in mm}

\item{introduce_NA}{TRUE or FALSE to simulate missing data. Introdcues random missings from 1 to 100 samples}

\item{NA_proportion}{proportion of data to set to missing}

\item{introduce_noise}{TRUE or FALSE to introduce data noise.}

\item{noise_magnitude}{magnitude of noise as fraction of standard deviation.}

\item{start_value_pupil_size}{pupil size start value in mm.}

\item{drift}{downward drift in pupil signal per response in mm}
}
\value{
sampled pupil data.
}
\description{
generates a vector in the shape of pupil data.
}
\examples{
#to be done
}
