% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pupil_data.R
\name{generate_pupil_data}
\alias{generate_pupil_data}
\title{generate_pupil_data}
\usage{
generate_pupil_data(
  sampling_rate = 300,
  phase_length = round(sampling_rate/2),
  number_of_responses = 100,
  mean_response_magnitude = 0.05,
  introduce_NA = T,
  NA_proportion = 0.2,
  start_value_pupil_size = 4,
  drift = -0.002
)
}
\arguments{
\item{sampling_rate}{sampling rate of the sampled data.}

\item{number_of_responses}{number of pupil responses to sample.}

\item{mean_response_magnitude}{mean pupil response magnitude in mm}

\item{introduce_NA}{simulate missing data}

\item{NA_proportion}{proportion of data to set to missing}

\item{start_value_pupil_size}{pupil size start value in mm.}

\item{drift}{drift in pupil signal per response in mm}

\item{phase_length_range}{range of pupil response durations to sample.}
}
\value{
sampled pupil data.
}
\description{
generates a vector in the shape of pupil data.
}
\examples{
#to be done
}
