% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PupilPreprocess-package.R
\docType{package}
\name{PupilPreprocess-package}
\alias{PupilPreprocess}
\alias{PupilPreprocess-package}
\title{PupilPreprocess: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicobast/PupilPreprocess}
  \item Report bugs at \url{https://github.com/nicobast/PupilPreprocess/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
